- Edit in patients_department_list_view file and make patients_department_bloc_builder main file - Add patients_department_bloc_builder - Add patients_department_list_view_item - Add patients_department_shimmer_loading
