+++ lib/feature/admin-home/ui/widgets/patients_department/patients_department_list_view_item.dart
@@ -1,0 +1,92 @@
+import 'package:flutter/material.dart';
+import 'package:leuko_care/core/resourses/colors_manager.dart';
+import 'package:leuko_care/core/resourses/fonts_manager.dart';
+import 'package:leuko_care/core/resourses/sizes_util_manager.dart';
+import 'package:leuko_care/core/resourses/styles_manager.dart';
+import 'package:leuko_care/feature/patients/data/models/patient_model.dart';
+
+class PatientsDepartmentListViewItem extends StatelessWidget {
+  final PatientModel patient;
+
+  const PatientsDepartmentListViewItem({super.key, required this.patient});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: EdgeInsets.only(bottom: HeightManager.h16),
+      child: Row(
+        children: [
+          ClipRRect(
+            borderRadius: BorderRadius.circular(20),
+            child: Image.network(
+              patient.profileImage,
+              width: WidthManager.w110,
+              height: HeightManager.h120,
+              fit: BoxFit.cover,
+            ),
+          ),
+          SizedBox(width: WidthManager.w16),
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  patient.name,
+                  style: getBoldTextStyle(
+                    fontSize: FontSizeManager.s18,
+                    color: ColorsManager.darkBlue,
+                  ),
+                  overflow: TextOverflow.ellipsis,
+                ),
+                SizedBox(height: HeightManager.h5),
+
+                // الهاتف مع أيقونة
+                Row(
+                  children: [
+                    Icon(
+                      Icons.phone,
+                      size: 16,
+                      color: ColorsManager.gray,
+                    ),
+                    SizedBox(width: 6),
+                    Text(
+                      patient.phone,
+                      style: getMediumTextStyle(
+                        fontSize: FontSizeManager.s12,
+                        color: ColorsManager.gray,
+                      ),
+                    ),
+                  ],
+                ),
+
+                SizedBox(height: HeightManager.h5),
+
+                // الإيميل مع أيقونة
+                Row(
+                  children: [
+                    Icon(
+                      Icons.email,
+                      size: 16,
+                      color: ColorsManager.gray,
+                    ),
+                    SizedBox(width: 6),
+                    Expanded(
+                      child: Text(
+                        patient.email,
+                        style: getMediumTextStyle(
+                          fontSize: FontSizeManager.s12,
+                          color: ColorsManager.gray,
+                        ),
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
